Fixed the following Chad's observations:
In DashboardAction.java you have a couple System.out.println statements that you may have used for debugging. those can be commented out or removed.

In DashboardVerify.java you have a sleep statement. did you need to pause it for something? If you do then try to use a wait statement instead of sleep.

In Login_StepDefs.java it looks like you recreated a step def that is already there. You can use the one below where you created yours. It does the exact same thing. we want to try to avoid re-creating the same statements like that.

I would get rid of the "Given I am a counselor" since your login statement (when) takes care of the role. The Given I am a counselor does no action so it's not useful in this situation. You could use either of these:
Given I am logged in as a counselor
Given I am logged into a school as with

Also, in your data, rtd1 is a district account, not a school.